// src/app/interfaces/Emplado-inter.ts
// Puedes ponerla en src/app/models/empleado-item.ts
export interface EmpleadoItem {
  id?: number;
  rfc?: string; RFC?: string;
  curp?: string; CURP?: string;
  primer_apellido?: string; primerApellido?: string;
  segundo_apellido?: string; segundoApellido?: string;
  nombre?: string;
  empleado?: string;
  nombreCompleto?: string;
}